flask-email-backend/requirements.txt
Flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==2.2.3
pyOpenSSL==24.0.0
